Authorization

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task-repository';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
 import { GetTaskFilterDto } from './dto/get-task-filter.dto';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -12,12 +13,12 @@ export class TasksService {
     @InjectRepository(TaskRepository) private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<any> {
-    const task = await this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<any> {
+    const task = await this.taskRepository.createTask(createTaskDto, user);
     return {
       task: {
         id: task.id,
@@ -29,24 +30,30 @@ export class TasksService {
     };
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!found) {
       throw new NotFoundException(`Task with Id "${id}" not found`);
     }
     return found;
   }
 
-  async deleteTaskById(id: number): Promise<any> {
-    const task = await this.taskRepository.delete(id);
+  async deleteTaskById(id: number, user: User): Promise<any> {
+    const task = await this.taskRepository.delete({ id, userId: user.id });
     if (task.affected === 0) {
       throw new NotFoundException(`Task with Id "${id}" not found`);
     }
     return { message: 'Task deleted successfully' };
   }
 
-  async updateTaskStatusById(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatusById(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
